Your new commit message without secret

+++ Smoking-Cessation-Support-Platform/src/main/java/com/smokingcessation/dto/SubscriptionRenewalDTO.java
@@ -7,7 +7,7 @@ import java.time.LocalDate;
 @Data
 public class SubscriptionRenewalDTO {
     @NotNull(message = "Plan ID is required")
-    private Integer planId;
+    private Long planId;
 
     private LocalDate renewalDate;
     private String paymentMethod;
