+++ Smoking-Cessation-Support-Platform/src/main/java/com/smokingcessation/repository/SubscriptionRepository.java
@@ -5,8 +5,8 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface SubscriptionRepository extends JpaRepository<Subscription, Integer> {
+public interface SubscriptionRepository extends JpaRepository<Subscription, Long> {
     List<Subscription> findByIsActiveTrue();
     List<Subscription> findByMember_MemberID(Long memberID);
-    List<Subscription> findByMembershipPlan_PlanID(Integer planID);
+    List<Subscription> findByMembershipPlan_PlanID(Long planID);
 }
